+++ src/Information/Information.cpp
@@ -81,6 +81,49 @@ void Information::Run()
         memNetwork->parameter_master_present = this->network->parameter_master_present;
     }
 
+    // -- Network WiFi
+    if (this->network->wifiConnected != memNetwork->wifiConnected ||
+        this->network->hostname != memNetwork->hostname ||
+        this->network->ipAddress != memNetwork->ipAddress ||
+        this->network->subnetmask != memNetwork->subnetmask ||
+        this->network->macAddress != memNetwork->macAddress ||
+        this->network->gateway != memNetwork->gateway )
+    {  
+        FormatPrintNetworkWiFi(BollToConnectionState(this->network->wifiConnected),
+                               this->network->hostname,
+                               this->network->ipAddress,
+                               this->network->subnetmask,
+                               this->network->macAddress,
+                               this->network->gateway);
+
+        memNetwork->wifiConnected = this->network->wifiConnected;
+        memNetwork->hostname = this->network->hostname;
+        memNetwork->ipAddress = this->network->ipAddress;
+        memNetwork->subnetmask = this->network->subnetmask;
+        memNetwork->macAddress = this->network->macAddress;
+        memNetwork->gateway = this->network->gateway;
+    }
+
+    // -- Network MQTT
+    if (this->network->mqttConnected != memNetwork->mqttConnected ||
+        this->network->clientState != memNetwork->clientState ||
+        this->network->brokerIpAddress != memNetwork->brokerIpAddress ||
+        this->network->brokerPort != memNetwork->brokerPort ||
+        this->network->clientName != memNetwork->clientName)
+    {
+        FormatPrintNetworkMQTT(BollToConnectionState(this->network->mqttConnected),
+                               IntToClientStateInfo(this->network->clientState),
+                               this->network->brokerIpAddress,
+                               this->network->brokerPort,
+                               this->network->clientName);
+
+        memNetwork->mqttConnected = this->network->mqttConnected;
+        memNetwork->clientState = this->network->clientState;
+        memNetwork->brokerIpAddress = this->network->brokerIpAddress;
+        memNetwork->brokerPort = this->network->brokerPort;
+        memNetwork->clientName = this->network->clientName;
+    }
+
     // -- Motion Parameter
     if (this->network->stNetworkMotionData.motionDetectionEnabled != memNetwork->stNetworkMotionData.motionDetectionEnabled ||
         this->network->stNetworkMotionData.timeBasedBrightnessChangeEnabled != memNetwork->stNetworkMotionData.timeBasedBrightnessChangeEnabled ||
@@ -544,6 +587,113 @@ void Information::FormatPrintMotionDetected(String motionDetected,
     BottomSpacerPrint();
 };
 
+
+/**
+ * Prints a network wifi formatted message to serial
+ * 
+ * @param connectionState   The current connection state of the WiFi connection
+ * @param hostname          The current hostname of the client if connected
+ * @param ipAddress         The current ip address of the esp if connected
+ * @param subnetmask        The current subnetmask of the esp if connected
+ * @param macAddress        The current macAddress of the esp
+ * @param gateway           The current gateway of the esp if connected
+ **/
+void Information::FormatPrintNetworkWiFi(String connectionState,
+                                         String hostname,
+                                         String ipAddress,
+                                         String subnetmask,
+                                         String macAddress,
+                                         String gateway)
+{
+    TopSpacerPrint();
+
+    // Network WiFi changed
+    InsertPrint();
+    Serial.println(F("Network WiFi changed"));
+
+    // Connection State
+    InsertPrint();
+    Serial.print(F("Connection State        : "));
+    Serial.println(connectionState);
+
+    // Hostname
+    InsertPrint();
+    Serial.print(F("Hostname                : "));
+    Serial.println(hostname);
+
+    // IP Address
+    InsertPrint();
+    Serial.print(F("IP Address              : "));
+    Serial.println(ipAddress);
+
+    // Subnetmask
+    InsertPrint();
+    Serial.print(F("Subnetmask              : "));
+    Serial.println(subnetmask);
+
+    // Mac Address
+    InsertPrint();
+    Serial.print(F("Mac Address             : "));
+    Serial.println(macAddress);
+
+    // gateway
+    InsertPrint();
+    Serial.print(F("Gateway                 : "));
+    Serial.println(gateway);
+
+    BottomSpacerPrint();
+};
+
+
+/**
+ * Prints a network mqtt formatted message to serial
+ * 
+ * @param connectionState   The current connection state of the mqtt connection
+ * @param clientStateInfo   The current client state info
+ * @param brokerIpAddress   The broker ip address to connect to
+ * @param clientName        The mqtt client name
+ **/
+void Information::FormatPrintNetworkMQTT(String connectionState,
+                                         String clientStateInfo,
+                                         String brokerIpAddress,
+                                         int brokerPort,
+                                         String clientName)
+{
+    TopSpacerPrint();
+
+    // Network MQTT changed
+    InsertPrint();
+    Serial.println(F("Network MQTT changed"));
+
+    // Connection State
+    InsertPrint();
+    Serial.print(F("Connection State        : "));
+    Serial.println(connectionState);
+
+    // Client State Info
+    InsertPrint();
+    Serial.print(F("Client State Info       : "));
+    Serial.println(clientStateInfo);
+
+    // Broker IP Address
+    InsertPrint();
+    Serial.print(F("Broker IP Address       : "));
+    Serial.println(brokerIpAddress);
+
+    // Broker Port
+    InsertPrint();
+    Serial.print(F("Broker Port             : "));
+    Serial.println(brokerPort);
+
+    // Client Name
+    InsertPrint();
+    Serial.print(F("Client Name             : "));
+    Serial.println(clientName);
+
+    BottomSpacerPrint();
+};
+
+
 /**
  * Prints a formatted top spacer message to serial
  **/
@@ -591,6 +741,75 @@ String Information::BoolToString(bool b)
         return "false";
     }
 };
+
+
+/**
+ * Converts a given bool value to the corresponding connection state
+ * 
+ * @param b     The bool value to convert
+ * 
+ * @return The value of the bool as connection state
+ **/
+String Information::BollToConnectionState(bool b)
+{
+    if (b)
+    {
+        return "Connected";
+    }
+    else
+    {
+        return "Disconnected";
+    }
+};
+
+
+/**
+ * Converts a given int value to the corresponding client state info
+ * 
+ * @param i     The int value to convert
+ * 
+ * @return The value of the int as client state info
+ **/
+String Information::IntToClientStateInfo(int i)
+{
+    switch(i)
+    {
+
+    case -4:
+        return "Connection Timeout";
+
+    case -3:
+        return "Connection Lost";
+
+    case -2:
+        return "Connect Failed";
+
+    case -1:
+        return "Disconnected";
+
+    case 0:
+        return "Connected";
+
+    case 1:
+        return "Connect Bad Protocol";
+
+    case 2:
+        return "Connect Bad Client ID";
+
+    case 3:
+        return "Connect Unavailable";
+
+    case 4:
+        return "Connect Bad Credentials";
+
+    case 5:
+        return "Connect Unauthorized";
+
+    default:
+        return "Unknown";
+    }
+};
+
 
 /**
  * Converts a SingleLEDEffect to a String